Modificación de entidades para implementar opciones en las relaciones. Creados DTOs cortos para su implementación en los DTO completos (ya realizado). Modificación de los servicios, para un mejor mapeo de los datos. Modificación de los tipos en los controladores para coincidir con los cambios realizados en los servicios. Comienzo de implementación de Spring Security (no definitivo).

+++ src/main/java/com/metaenlace/CitasMedicas/Services/CitaServiceImpl.java
@@ -3,27 +3,25 @@ package com.metaenlace.CitasMedicas.Services;
 import com.metaenlace.CitasMedicas.DTO.CitaDTO;
 import com.metaenlace.CitasMedicas.Entities.Cita;
 import com.metaenlace.CitasMedicas.Repositories.CitaRepository;
+import com.metaenlace.CitasMedicas.Repositories.DiagnosticoRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.modelmapper.ModelMapper;
 
+import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class CitaServiceImpl implements CitaService{
     @Autowired
     private CitaRepository citaRepository;
+    @Autowired
+    private DiagnosticoRepository diagnosticoRepository;
+
     private final ModelMapper mapper = new ModelMapper();
 
-    private CitaDTO mapToDTO(Cita obj){
-        CitaDTO dto = mapper.map(obj, CitaDTO.class);
-        return dto;
-    }
-    private CitaDTO mapToEntity(CitaDTO obj){
+    private CitaDTO convertDTO(Cita obj){
         CitaDTO dto = mapper.map(obj, CitaDTO.class);
         return dto;
     }
@@ -33,31 +31,36 @@ public class CitaServiceImpl implements CitaService{
     @Transactional(readOnly = true)
     public List<CitaDTO> listadoCitas() {
         List<Cita> citas = citaRepository.findAll();
-        return citas.stream().map(cita -> new CitaDTO(cita.getId(), cita.getFechaHora(), cita.getMotivoCita(), cita.getAttribute11(), cita.getDiagnostico())).collect(Collectors.toList());
+        List<CitaDTO> citaDTOList = new LinkedList<CitaDTO>();
+        for (Cita cita: citas){
+            citaDTOList.add(convertDTO(cita));
+        }
+        return citaDTOList;
     }
 
     @Override
     public CitaDTO findCitaById(long id) {
-        Optional<Cita> citaPorId = citaRepository.findById(id);
-        return citaPorId.map(cita -> new CitaDTO(cita.getId(),cita.getFechaHora(), cita.getMotivoCita(), cita.getAttribute11(), cita.getDiagnostico())).orElse(null);
+       Cita citaPorId = citaRepository.findById(id).orElse(null);
+        if (citaPorId == null){
+            throw new RuntimeException("Cita no encontrada");
+        } else {
+            return convertDTO(citaPorId);
+        }
     }
 
-    /*
-    @Override
-    @Transactional(readOnly = true)
-    public List<Cita> listadoPorFechaHora(Date fh) {
-        return citaRepository.findbyFechaHora(fh);
-    }
-    */
     @Override
     @Transactional
     public void deleteCita(long id) {
-        citaRepository.deleteById(id);
+        if (citaRepository.existsById(id)) {
+            citaRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Cita no encontrada");
+        }
     }
 
     @Override
-    @Transactional
-    public void saveCita(CitaDTO cita) {
-        citaRepository.save(mapToEntity(cita));
+    public CitaDTO saveCita(Cita cita) {
+        citaRepository.save(cita);
+        return convertDTO(cita);
     }
 }
