+++ src/main/java/com/metaenlace/CitasMedicas/Services/DiagnosticoServiceImpl.java
@@ -3,13 +3,13 @@ package com.metaenlace.CitasMedicas.Services;
 import com.metaenlace.CitasMedicas.DTO.DiagnosticoDTO;
 import com.metaenlace.CitasMedicas.Entities.Diagnostico;
 import com.metaenlace.CitasMedicas.Repositories.DiagnosticoRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.LinkedList;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class DiagnosticoServiceImpl implements DiagnosticoService{
@@ -17,30 +17,49 @@ public class DiagnosticoServiceImpl implements DiagnosticoService{
     @Autowired
     private DiagnosticoRepository diagnosticoRepository;
 
+    private final ModelMapper mapper = new ModelMapper();
+
+    private DiagnosticoDTO convertDTO(Diagnostico diag){
+        return mapper.map(diag, DiagnosticoDTO.class);
+    }
+
     @Override
     @Transactional(readOnly = true)
     public List<DiagnosticoDTO> listadoDiagnosticos() {
         List<Diagnostico> diagnosticos = diagnosticoRepository.findAll();
-        return diagnosticos.stream().map(diagnostico -> new DiagnosticoDTO(diagnostico.getId(), diagnostico.getValoracionEspecialista(), diagnostico.getEnfermedad())).collect(Collectors.toList());
+        List<DiagnosticoDTO> diagnosticoDTOList = new LinkedList<>();
+        for (Diagnostico diag: diagnosticos) {
+            diagnosticoDTOList.add(convertDTO(diag));
+        }
+        return diagnosticoDTOList;
     }
 
     @Override
     @Transactional(readOnly = true)
     public DiagnosticoDTO findDiagById(long id) {
-        Optional<Diagnostico> diag = diagnosticoRepository.findById(id);
-        return diag.map(diagnostico -> new DiagnosticoDTO(diagnostico.getId(), diagnostico.getValoracionEspecialista(), diagnostico.getEnfermedad())).orElse(null);
+        Diagnostico diag = diagnosticoRepository.findById(id).orElse(null);
+        if (diag == null) {
+            throw new RuntimeException("Diagnostico no encontrado");
+        } else {
+            return convertDTO(diag);
+        }
     }
 
     @Override
     @Transactional
     public void deleteDiag(long id) {
-        diagnosticoRepository.deleteById(id);
+        if (diagnosticoRepository.existsById(id)) {
+            diagnosticoRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Diagnostico no encontrado");
+        }
+
     }
 
     @Override
-    @Transactional
-    public void saveDiag(DiagnosticoDTO diag)
+    public DiagnosticoDTO saveDiag(Diagnostico diag)
     {
         diagnosticoRepository.save(diag);
+        return convertDTO(diag);
     }
 }
